Added fragments for quiz questions in portrait and landscape mode. Triggers autmotically, because of placement in layout for portrait and layout-land for landscape
